Resolved all issues with shutdown and refactoring
